requirements.txt
numpy
pandas
pyvista
torch
pytest
flake8
tqdm
pyyaml